replace size_t with uint32_t for diagnostic storage

+++ src/tracing/esp32_diagnostic_trace/DiagnosticEntry.h
@@ -23,7 +23,7 @@ namespace chip {
 namespace Tracing {
 namespace Diagnostics {
 
-static constexpr size_t kMaxStringValueSize = 64;
+static constexpr uint32_t kMaxStringValueSize = 64;
 
 enum class ValueType
 {
